Made Enemy a scriptable object, fixed a collision bug

+++ X_Breach/Assets/Scripts/BaseEntity.cs
@@ -51,7 +51,7 @@ public class BaseEntity
 
     public void CauseDamage(BaseEntity other) { other.TakeDamage(damage); }
 
-    public void TakeDamage(float dmg)
+    public void TakeDamage(int dmg)
     {
         if (!IsDead())
             health -= dmg;
